requirements.txt
Telethon~=1.35.0
PyMySQL~=1.1.0
python-telegram-bot~=21.2
python-binance
mplfinance
pandas